requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
faiss-cpu
pydantic
waitress